#37 fixed the remarks
the module was renamed, arguments were fixed, brackets were removed
